+++ tcob/src/app/Platform.cpp
@@ -81,7 +81,7 @@ platform::platform(bool headless, game::init const& ginit)
     register_service<task_manager>(std::make_shared<task_manager>(
         ginit.WorkerThreads
             ? *ginit.WorkerThreads
-            : static_cast<i32>(std::thread::hardware_concurrency() * 2)));
+            : static_cast<isize>(std::thread::hardware_concurrency() * 2)));
 
     // init config formats
     InitConfigFormats();
